When raw_body is provided always use it.
According to Plug.Conn docs body cannot be fetched multiple times,
so if we have a raw_body is reasonable that someone already fetched it.
In this case we *must* avoid reading it again, because it can stall the connection
since Plug adapter will wait until timeout for a never arriving additional body data.

Also docs explicitly states that if raw_body is present is always used,
so let's enforce such case.