fix: how host is being read and set from Plug.Conn
Plug.Conn has a special key host that is used to store host information instead
of relying on the list of headers. In newer versions of Plug.Conn, it raises an
error if put_req_header is used to set host to prevent this mistake.

Read host from the right place and update tests.